Fixed network bug caused by the use of the ready() method in BufferedReader . All network messages are now newline delimited (have a "\n" added at the end), and all reads over the network are now done with "readLine()".
As it turns out, the bug in which the online players list would "disappear" then reappear, the bug where the server would occasionally "smush" two messages together (which we "fixed" by adding spaces) and the bug where the two Game instances would sometimes lose the end of a message string, were all related to one issue. 

Our program used BufferedReader's "ready()" method in it's loop to determine if the socket's stream was "ready" to read additional chars. However, ready()=true guarantees that the next call of read() will NOT block. However, read() = false does NOT guarantee that the next call to read() would have blocked.

TCP is a stream-oriented protocol. That means that for every string of bytes you send it, TCP does not guarantee the segments that it will send the bytes in at an OS level. For maximum efficiency, the TCP stack on the operating system looks at the message that needs to be sent, then breaks it up into chunks to be reassambled at the other end.

Somehow, when the low-level TCP stack broke up the message, the "ready()" method in Java's buffered reader got spooked that the Stream might block. Thus, ready() returned false early, and the message was broken up until the next send, where it would be tagged onto the next message. Getting rid of the use of the "ready()" method in reading messages, and transitioning to a newline delimited system fixed the problem.